requirements.txt
droidrun[google]
llama-index-llms-google-genai
google-generativeai
adbutils
